requirements.txt
Django
djangorestframework
mysqlclient
django-cors-headers
pika
django-oauth-toolkit
django-rest-knox
pillow == 9.0